memory_agent/requirements.txt
annotated-types==0.7.0
anyio==4.12.1
backoff==2.2.1
certifi==2026.1.4
charset-normalizer==3.4.4
colorama==0.4.6
distro==1.9.0
fsspec==2024.12.0
greenlet==3.3.1
grpcio==1.76.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
jiter==0.13.0
json_repair==0.44.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain-classic==1.0.1
langchain-core==1.2.8
langchain-neo4j==0.8.0
langchain-text-splitters==1.1.0
langgraph==1.0.7
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.7
langgraph-sdk==0.3.3
langsmith==0.6.8
mem0ai==1.0.3
neo4j==5.28.3
neo4j-graphrag==1.13.0
numpy==2.4.2
openai==2.16.0
orjson==3.11.7
ormsgpack==1.12.2
packaging==26.0
portalocker==3.2.0
posthog==7.8.0
protobuf==5.29.5
pydantic==2.12.5
pydantic_core==2.41.5
pypdf==6.6.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
pywin32==311
PyYAML==6.0.3
qdrant-client==1.16.2
rank-bm25==0.2.2
requests==2.32.5
requests-toolbelt==1.0.0
scipy==1.17.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.46
tenacity==9.1.2
tqdm==4.67.2
types-PyYAML==6.0.12.20250915
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.0
xxhash==3.6.0
zstandard==0.25.0
